requirements.txt
selenium==4.24.0
flask==3.0.0
sqlalchemy>=2.0.16
python-dotenv==1.0.0
webdriver-manager==4.0.0
rich==13.7.0
flask-sqlalchemy==3.1.1
